An critical subtlety had been overlooked in the design of the structural analysis algorithm when a region is identified and reduced. Originally, all incoming/outgoing edges to/from the region were removed and then inserted into the new abstract node. However, this alters the order of the edges as they appear in the successor/predecessor lists of the nodes outside of the region.
This order is important, for example, in the identification of 'if-then-else'
regions to determine which is the 'then' and which is the 'else' block. Figuring
this out at the time of code emission is very tough. But this information is
available in the control tree since its original construction from the
control-flow graph ('then' block will appear always before the 'else' block in
the basic block's successor list). We just need to be careful to keep this
information intact during the control tree manipulation.

The solution has been reconnecting nodes (functions 'cnode_reconnect_src' and
'cnode_reconnect_dest') instead of removing them and adding them again
(functions 'cnode_disconnect' and 'cnode_connect').